Construct new edges before starting to create new faces.

+++ client/js/editor/editor.js
@@ -349,7 +349,7 @@ shapy.editor.Editor.prototype.setCanvas = function(canvas) {
 
   // Initialise the layout.
   this.vp_.width = this.vp_.height = 0;
-  this.layout = new shapy.editor.Layout.Double(this);
+  this.layout = new shapy.editor.Layout.Single(this);
   this.rig(this.rigTranslate_);
 
   // Start checkpointing.
